Add new types to Analog temp and fix TMP37

+++ Source/Meadow.Foundation.Core/Sensors/Temperature/AnalogTemperature.Configuration.cs
@@ -16,20 +16,20 @@ namespace Meadow.Foundation.Sensors.Temperature
             ///     Sample reading as specified in the product data sheet.
             ///     Measured in degrees Centigrade.
             /// </summary>
-            public int SampleReading { get; protected set; } = 25;
+            public double SampleReading { get; protected set; } = 25;
 
             /// <summary>
             ///     Millivolt reading the sensor will generate when the sensor
             ///     is at the Samplereading temperature.  This value can be
             ///     obtained from the data sheet. 
             /// </summary>
-            public int MillivoltsAtSampleReading { get; protected set; } = 250;
+            public double MillivoltsAtSampleReading { get; protected set; } = 250;
 
             /// <summary>
             ///     Linear change in the sensor output (in millivolts) per 1 degree C
             ///     change in temperature.
             /// </summary>
-            public int MillivoltsPerDegreeCentigrade { get; protected set; } = 10;
+            public double MillivoltsPerDegreeCentigrade { get; protected set; } = 10;
 
             /// <summary>
             ///     Default constructor.  Create a new Calibration object with default values
@@ -46,9 +46,9 @@ namespace Meadow.Foundation.Sensors.Temperature
             /// <param name="millivoltsAtSampleReading">Millivolts output at the sample reading (from the data sheet).</param>
             /// <param name="millivoltsPerDegreeCentigrade">Millivolt change per degree centigrade (from the data sheet).</param>
             /// <param name="millivoltsOffset">Millovolts offset (from the data sheet).</param>
-            public Calibration(int degreesCelciusSampleReading,
-                               int millivoltsAtSampleReading,
-                               int millivoltsPerDegreeCentigrade)
+            public Calibration(double degreesCelciusSampleReading,
+                               double millivoltsAtSampleReading,
+                               double millivoltsPerDegreeCentigrade)
             {
                 SampleReading = degreesCelciusSampleReading;
                 MillivoltsAtSampleReading = millivoltsAtSampleReading;
